CSRF tokens, Server side price validation, quantity validation

+++ delete.php
@@ -6,32 +6,34 @@
 <?php include("./includes/header.inc.php");?><br>
 <?php
 
-$myid = strip_tags($_REQUEST['id']);
+if($_SESSION['csrf_token']===$_GET['csrf_token']){
+	if(strip_tags($_SESSION['username'])) {
 
-if(strip_tags($_SESSION['username'])) {
-	$sql = "DELETE FROM products WHERE id=?";
+		$myid = strip_tags($_GET['id']);
+		$sql = "DELETE FROM products WHERE id=?";
 
-	//Create a prepared statement
-	$stmt = mysqli_stmt_init($mysqli);
+		//Create a prepared statement
+		$stmt = mysqli_stmt_init($mysqli);
 
-	//Prepare the statement
-	if (!mysqli_stmt_prepare($stmt, $sql)){
-		echo "SQL statement preperation failed:" . $mysqli->error;
-	} else {
-		// Bind the parameters to the placeholder values in the sql variable query
-		mysqli_stmt_bind_param($stmt, "s", $myid);
-		
-		// Execute the statement
-		if (mysqli_stmt_execute($stmt)) {
-			echo "Product $myname deleted.";
+		//Prepare the statement
+		if (!mysqli_stmt_prepare($stmt, $sql)){
+			echo "Error T-T";
 		} else {
-			"Execution failed: " . mysqli_stmt_error($stmt);
+			// Bind the parameters to the placeholder values in the sql variable query
+			mysqli_stmt_bind_param($stmt, "s", $myid);
+			
+			// Execute the statement
+			if (mysqli_stmt_execute($stmt)) {
+				echo "Product $myname deleted.";
+			} else {
+				echo "Error T-T";
+			}
 		}
+	} else{
+		echo "Slow down there hacker... You aren't allowed to do that without logging in <.<";
 	}
-
-
 } else {
-	echo "Slow down there hacker... You aren't allowed to do that without logging in <.<";
+	echo "Error T-T";
 }
